requirements.txt
Flask==2.0.1
tensorflow==2.13.0
Pillow==9.5.0
numpy==1.24.3
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0 